+++ missiongen/__init__.py
@@ -1,4 +1,4 @@
-__version__ = "1.6.2"
+__version__ = "1.6.3"
 
 from .recipe import Recipe
 from .builder import StarterBuilder, generate